+++ Sources/code/platform/Models/Content/OperationResponseSchemaContentPlatformModel.swift
@@ -5,23 +5,29 @@ import Foundation
 
 public extension PlatformClient.Content {
     /*
-        Model: DeletedResource
+        Model: OperationResponseSchema
         Used By: Content
     */
 
-    class DeletedResource: Codable {
+    class OperationResponseSchema: Codable {
         
+        
+        public var success: Bool
         
         public var message: String?
         
 
         public enum CodingKeys: String, CodingKey {
             
+            case success = "success"
+            
             case message = "message"
             
         }
 
-        public init(message: String? = nil) {
+        public init(message: String? = nil, success: Bool) {
+            
+            self.success = success
             
             self.message = message
             
@@ -29,6 +35,11 @@ public extension PlatformClient.Content {
 
         required public init(from decoder: Decoder) throws {
             let container = try decoder.container(keyedBy: CodingKeys.self)
+            
+            
+                success = try container.decode(Bool.self, forKey: .success)
+                
+            
             
             
                 do {
@@ -46,6 +57,11 @@ public extension PlatformClient.Content {
         
         public func encode(to encoder: Encoder) throws {
             var container = encoder.container(keyedBy: CodingKeys.self)
+            
+            
+            
+            try? container.encodeIfPresent(success, forKey: .success)
+            
             
             
             
@@ -61,23 +77,29 @@ public extension PlatformClient.Content {
 
 public extension PlatformClient.ApplicationClient.Content {
     /*
-        Model: DeletedResource
+        Model: OperationResponseSchema
         Used By: Content
     */
 
-    class DeletedResource: Codable {
+    class OperationResponseSchema: Codable {
         
+        
+        public var success: Bool
         
         public var message: String?
         
 
         public enum CodingKeys: String, CodingKey {
             
+            case success = "success"
+            
             case message = "message"
             
         }
 
-        public init(message: String? = nil) {
+        public init(message: String? = nil, success: Bool) {
+            
+            self.success = success
             
             self.message = message
             
@@ -85,6 +107,11 @@ public extension PlatformClient.ApplicationClient.Content {
 
         required public init(from decoder: Decoder) throws {
             let container = try decoder.container(keyedBy: CodingKeys.self)
+            
+            
+                success = try container.decode(Bool.self, forKey: .success)
+                
+            
             
             
                 do {
@@ -102,6 +129,11 @@ public extension PlatformClient.ApplicationClient.Content {
         
         public func encode(to encoder: Encoder) throws {
             var container = encoder.container(keyedBy: CodingKeys.self)
+            
+            
+            
+            try? container.encodeIfPresent(success, forKey: .success)
+